Add initial versions of the server side scripts
The slurm_management_daemon runs an infinit loop checking the current users directory for projects and then calls the
poll_and_launch_project script with this project directory. That script then pulls from the barebone git repository for this
user and project and checks for the submit.sh file. If it finds that file, it will launch that file through slurm and write back
the jobid file into crun.jobid.